Split SettingsContent.kt into one file per settings category
The 2280-line file mixed all eight Settings categories (Setup, UI,
Widgets, Video Playback, Other, App Drawer, Sound, Thor) plus shared
chrome (SettingsLabel, category rows) in one file. Split along the
existing SettingsCategory boundaries, with a SettingsShared.kt for
pieces used across more than one category (SettingsLabel,
SettingsItemShape, ToggleSettingRow, etc). No behavior change -
pure move, updated config/detekt/baseline.xml entries to follow
their functions to the new files.
