+++ app/src/main/java/com/esde/companion/ui/settings/WidgetsSettingsContent.kt
@@ -1,0 +1,68 @@
+package com.esde.companion.ui.settings
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ChevronRight
+import androidx.compose.material.icons.filled.Widgets
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+/**
+ * Widgets category content - just hands off to the full-screen widget editor, same
+ * "navigate out" idiom Edit Widgets used under UI Settings before it got its own
+ * top-level category.
+ */
+@Composable
+internal fun WidgetsSettingsContent(onEditWidgetsClick: () -> Unit) {
+    Column(
+        modifier =
+            Modifier
+                .fillMaxSize()
+                .verticalScroll(rememberScrollState())
+                .padding(24.dp),
+        verticalArrangement = Arrangement.spacedBy(24.dp),
+    ) {
+        EditWidgetsEntry(onClick = onEditWidgetsClick)
+    }
+}
+
+@Composable
+private fun EditWidgetsEntry(onClick: () -> Unit) {
+    Surface(
+        onClick = onClick,
+        modifier = Modifier.fillMaxWidth(),
+        shape = SettingsItemShape,
+        color = MaterialTheme.colorScheme.surfaceContainerHigh.copy(alpha = SETTINGS_PANEL_ALPHA),
+    ) {
+        Row(
+            modifier =
+                Modifier
+                    .fillMaxWidth()
+                    .padding(16.dp),
+            horizontalArrangement = Arrangement.SpaceBetween,
+            verticalAlignment = Alignment.CenterVertically,
+        ) {
+            Column(verticalArrangement = Arrangement.spacedBy(4.dp)) {
+                SettingsLabel(icon = Icons.Filled.Widgets, text = "Edit Widgets")
+                Text(
+                    text = "Add, move, and resize widgets on the main screen",
+                    style = MaterialTheme.typography.bodySmall,
+                )
+            }
+            Icon(imageVector = Icons.Filled.ChevronRight, contentDescription = null)
+        }
+    }
+}
